+++ app/__init__.py
@@ -1,0 +1,7 @@
+from flask import Flask
+
+def create_app():
+    app = Flask(__name__)
+    app.config['SECRET_KEY'] = 'abcdefg7749177013'
+    
+    return app
