--- app/static/css/style2.css
@@ -1,111 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@500&display=swap');
-
-*{
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-}
-
-body {
-    margin: 0;
-    height: 100vh; /* 100% of viewport height */
-    display: flex;
-    background-color: #000000;
-}
-
-.Main-container {
-    font-family: 'Roboto', sans-serif;
-    font-weight: 500;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    align-items: center ;
-    flex-direction: column;
-}
-
-.img-title {
-    width: 30%;
-    height: auto;
-    max-width: 100%;
-    max-height: 100%;
-}
-
-@media (max-width: 768px) {
-    .img-title {
-        width: 100%;
-    }
-}
-
-.input {
-    width: 80%;
-    height: 20%;
-    flex: 1;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-}
-
-.search_bar {
-    margin-top: 40px;
-    display: flex;
-    align-items: center;
-    border-bottom: .1rem solid white;
-}
-
-.search_bar input {
-    font-size: medium;
-    background: none;
-    flex: 1;
-    border: 0;
-    outline: none;
-    color: white;
-    padding: 12px 8px;
-}
-
-#reset {
-    background-color: black;
-    border: 0;
-    cursor: pointer;
-    padding: 5px 8px;
-}
-
-#reset img {
-    background-color: white;
-    height: 50%;
-    width: 50%;
-}
-
-.extract {
-    background-color: black;
-    border: none;
-    display: inline-block;
-    cursor: pointer;
-}
-
-.extract img {
-    height: 20%;
-    width: 20%;
-}
-
-.vid-info {
-    border : solid 2px white;
-    border-radius: 10px;
-    color: #ffffff;
-    width: 60%;
-    margin-bottom: 500px;
-}
-
-.vid-general {
-    color: white;
-    padding: 7px 14px;
-    border-right: solid 1px white;
-    width: 50%;
-    float: left;
-}
-
-.vid-quality {
-    padding: 7px 14px;
-    border-left: solid 1px white;
-    width: 50%;
-    float: right;
-}
