+++ app/templates/index.html
@@ -4,30 +4,27 @@
         <title>YouTube Downloader</title>
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="icon" href="../static/images/favicon.ico" type="image/x-icon" sizes="20x20">
+        <link rel="icon" href="../static/images/favicon.ico" type="image/x-icon" sizes="32x32">
         <link rel="stylesheet" type="text/css" href="../static/css/style.css" />
         <script src="../static/js/script.js"></script>
     </head>
     <body>
-        <header>
-            <img class="logo" src="../static/images/logo.png" href="/" alt="logo">
-            <nav>
-                <ul class="nav__links">
-                    <li><a href="#">Home</a></li>
-                    <li><a href="#">About</a></li>
-                </ul>
-            </nav>
-            <a class="cta" href="login"><button type="button" class="SignButton">Sign In / Sign Up</button></a>
-        </header>   
-        <body>
-            <div class="container">   
-                <h1 class="title">YouTube Downloader</h1>
-                <form id="download-form" method="POST" class="search-bar"> 
-                    <input type="text" autocomplete="off" name="url" id="url" placeholder="Search video or Insert URL"/>
-                    <button onclick="setAction('/dl')" style="background: transparent;"><img src="../static/images/download.png"></button>
-                    <button onclick="setAction('/extract')" style="background: transparent;"><img src="../static/images/extract.png"></button>
-                </form>
-            </div>  
-        </body>
-    </body> 
-</html>     
+        <div class = "Main-container">
+            <img class="img-title" src="../static/images/title.png" href="/" alt="logo">
+            <form id="download-form" class="input" method="POST"> 
+                <div class="search_bar">
+                    <input type="text" autocomplete="off" name="url" id="url" placeholder="URL goes here"/>
+                    <button type="reset" id="reset" onclick="reset()"><img src="../static/images/clear.png"></button>
+                </div>
+                <button class="extract" onclick="setAction('/ytdl/extract')"><img src="../static/images/extract.png" alt="extract"></button>
+            </form>
+            <!-- <div class="vid-info"> 
+                <div class="vid-general" id="vid-general-info"></div>  
+                <div class="vid-quality">
+                    <p>Quality: </p>
+                </div>
+            </div>
+            <div style="clear: both;"></div> -->
+        </div>
+    </body>
+</html>
