--- app/templates/index2.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>YouTube Downloader</title>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="icon" href="../static/images/favicon.ico" type="image/x-icon" sizes="20x20">
-        <link rel="stylesheet" type="text/css" href="../static/css/style.css" />
-        <script src="../static/js/script.js"></script>
-    </head>
-    <body>
-        <header>
-            <img class="logo" src="../static/images/logo.png" href="/" alt="logo">
-            <nav>
-                <ul class="nav__links">
-                    <li><a href="#">Home</a></li>
-                    <li><a href="#">About</a></li>
-                </ul>
-            </nav>
-            <a class="cta" href="login"><button type="button" class="SignButton">Sign In / Sign Up</button></a>
-        </header>   
-        <body>
-            <div class="container">   
-                <h1 class="title">YouTube Downloader</h1>
-                <form id="download-form" class="search-bar"> 
-                    <input type="text" autocomplete="off" id="url" placeholder="Search video or Insert URL"/>
-                    <button onclick="setAction('/dl')" style="background: transparent;"><img src="../static/images/download.png"></button>
-                    <button onclick="extractVideoInfo()" style="background: transparent;"><img src="../static/images/extract.png"></button>
-                </form>
-                <div id="videoInfoDisplay"></div>
-            </div>  
-        </body>
-    </body> 
-</html>
