--- app/templates/index3.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>YouTube Downloader</title>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="icon" href="../static/images/favicon.ico" type="image/x-icon" sizes="32x32">
-        <link rel="stylesheet" type="text/css" href="../static/css/style2.css" />
-        <script src="../static/js/script.js"></script>
-    </head>
-    <body>
-        <div class = "Main-container">
-            <img class="img-title" src="../static/images/title.png" href="/" alt="logo">
-            <form id="download-form" class="input" method="POST"> 
-                <div class="search_bar">
-                    <input type="text" autocomplete="off" name="url" id="url" placeholder="URL goes here"/>
-                    <button type="reset" id="reset" onclick="reset()"><img src="../static/images/clear.png"></button>
-                </div>
-                <button class="extract" onclick="setAction('/extract')"><img src="../static/images/extract.png" alt="extract"></button>
-            </form>
-            <!-- <div class="vid-info"> 
-                <div class="vid-general" id="vid-general-info"></div>  
-                <div class="vid-quality">
-                    <p>Quality: </p>
-                </div>
-            </div>
-            <div style="clear: both;"></div> -->
-        </div>
-    </body>
-</html>
